Make 'backspace_and_resume_hint' binding more specific.
Otherwise, the delete_pair binding gets overridden. This change assumes
that we don't want to resume hinting after deleting a pair. Granted, I
am not familiar with what backspace_and_resume_hint does, so that may
not be a correct assumption.
To facilitate this feature, add Conditions for being preceeded by and
followed by any of the pairs [], {}, (), '', and "" (triple-quotes are
not included).